requirements.txt
arrow==1.1.1
cachelib==0.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
docopt==0.6.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Session==0.4.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.25
numpy==1.21.0
pipreqs==0.4.11
protobuf==3.16.0
pyarrow==4.0.1
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
tqdm==4.61.2
urllib3==1.26.4
Werkzeug==1.0.1
yarg==0.1.9
